iOS: Email Class Created
